use int

+++ MultiPlexer_ADS1115/MultiPlexer_ADS1115.h
@@ -9,36 +9,36 @@
 #include <ADS1115_WE.h>
 
 struct ADS1115Result {
-  float channel0;
-  float channel1;
-  float channel2;
-  float channel3;
+  int16_t channel0;
+  int16_t channel1;
+  int16_t channel2;
+  int16_t channel3;
 };
 
 class MultiPlexer_ADS1115
 {
   public:
     MultiPlexer_ADS1115(
-      float max_voltage,
-      int range_min = 0,
-      int range_max = 127,
+      int16_t max_voltage,
+      int16_t range_min = 0,
+      int16_t range_max = 127,
       int i2c_addr = 0x48,
       ADS1115_RANGE voltage_range = ADS1115_RANGE_6144
     );
     void loop();
     void begin();
     ADS1115Result readAll();
-    float readChannel(ADS1115_MUX channel);
-    float readChannel0();
-    float readChannel1();
-    float readChannel2();
-    float readChannel3();
+    int16_t readChannel(ADS1115_MUX channel);
+    int16_t readChannel0();
+    int16_t readChannel1();
+    int16_t readChannel2();
+    int16_t readChannel3();
 
   private:
     int _i2cAddress;
-    int _maxVoltage;
-    int _rangeMin;
-    int _rangeMax;
+    int16_t _maxVoltage;
+    int16_t _rangeMin;
+    int16_t _rangeMax;
     ADS1115_WE* _adc;
     ADS1115_RANGE _voltageRange;
 };
